vis_marine_env/requirements.txt
basemap==1.4.1
basemap_data==1.3.2
imageio==2.33.1
matplotlib==3.7.2
netCDF4==1.6.2
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1
seaborn==0.13.2
tensorboard==2.12.1
pillow==10.4.0

